tests/wpt/web-platform-tests/css/work-in-progress/opera/animations/keyframe_selector_list.html

<!DOCTYPE html>
<html>
<head>
   <title>CSS Animations - comma-separated list in keyframes selector</title>
   <link rel="author" title="Opera Software ASA" href="http://www.opera.com/">
   <link rel="help" href="https://drafts.csswg.org/css3-animations/#keyframes">
   <meta name="flags" content="dom">
   <meta name="assert" content="The keyframe selector for a keyframe style rule
   consists of a comma-separated list of percentage values or the keywords
   'from' or 'to'">
   <script src="/resources/testharness.js"></script>
   <script src="/resources/testharnessreport.js"></script>
   <style type="text/css">
      @keyframes step-right {
         from, 33% {
            left: 0px;
         }
         34%, 66% {
            left: 50px;
         }
         67%, to {
            left: 100px;
         }
      }
      div {
         width: 100px;
         height: 100px;
         background: blue;
         position: relative;
      }
      .test {
         animation: step-right 1s forwards;
      }
   </style>
   <script type="text/javascript">
      setup({explicit_done:true});
      onload = function() {
         var testName = document.querySelector('title').firstChild.data;
         var div = document.querySelector('div');
         var divStyle = getComputedStyle(div, null);
         var log = document.querySelectorAll('p')[1].firstChild;
         var startOK = false;
         var middleOK = false;
         var timeout = setTimeout(function() {
            test(function(){assert_true(false);}, testName); done();
            log.data = 'FAIL';
         }, 1500);
         div.addEventListener('animationstart', function() {
            startOK = divStyle.left == '0px';
         });
         setTimeout(function() {
            middleOK = divStyle.left == '50px';
         }, 500);
         div.addEventListener('animationend', function() {
            clearTimeout(timeout);
            var passed = startOK && middleOK && divStyle.left == '100px';
            test(function(){assert_true(passed);}, testName); done();
            log.data = passed ? 'PASS if the animation occurred.' : 'FAIL';
         });
         div.className = 'test';
      }
   </script>
   <style>iframe{display:none}</style>
</head>
<body>
   <p>The square below should move right in two steps.</p>
   <p>Testing...</p>
   <div></div>
</body>
</html>
